requirements.txt
torch
tqdm
pandas
torchaudio
hydra-core
transformers
datasets